Add longuest activities cleaner.
Takes in entry the logfile or by default the latest modified, then calls
`stats.py` requesting for longuest sessions in json format. Uses `jq` to
parse the result, then using `dialog` asks you for the activities to be
replaced. Select the actual activity duration, it will add a Desktop
entry according to the input.